requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
chunkr-ai==0.0.49
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dataclasses-json==0.6.7
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
eval-type-backport==0.2.2
executing==2.2.0
fastapi==0.115.12
frozenlist==1.6.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gotrue==2.11.4
graphviz==0.20.3
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
ipython==9.2.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.3
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langgraph==0.4.2
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.42
lz4==4.4.4
marshmallow==3.26.1
matplotlib-inline==0.1.7
mistralai==1.7.0
multidict==6.4.3
mypy-extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.2.5
openai==1.77.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
playwright==1.52.0
pluggy==1.5.0
postgrest==1.0.1
prompt-toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
protoc-gen-openapiv2==0.0.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pyee==13.0.0
pygments==2.19.1
pyjwt==2.9.0
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyyaml==6.0.2
realtime==2.4.3
redis==5.3.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
storage3==0.11.3
strenum==0.4.15
supabase==2.15.1
supafunc==0.9.4
tavily-python==0.7.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
websockets==14.2
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
